+++ changelog.md
@@ -1,3 +1,8 @@
+# 0.3
+- Added the ability to recognize the system language.
+- Fixed the pseudo-random number generator.
+- New functions added.
+
 # 0.2
 - Added multithreading support.
 - Added new functions for working with the file system.